requirements.txt
git+https://github.com/kaymccormick/marshmallow.git
plaster_pastedeploy
pyramid >= 1.9a
pyramid_jinja2
pyramid_retry
pyramid_tm
waitress
zope.interface
zope.component
